Enable both Date#to_era and Time#to_era by just requiring 'era_ja'
To do so, move main logic to EraJa::Conversion module.
